InsightAI/requirements.txt
semantic-kernel
chromadb
sentence-transformers
rank-bm25
python-dotenv
pandas
openai
markdown
playwright
uvicorn
fastapi